+++ runtime/doc/news.txt
@@ -327,6 +327,8 @@ These existing features changed their behavior.
 • |$VIM| and |$VIMRUNTIME| no longer check for Vim version-specific runtime
   directory `vim{number}` (e.g. `vim82`).
 • 'scrollback' maximum value increased from 100000 to 1000000
+• |matchfuzzy()| and |matchfuzzypos()| use an improved fuzzy matching algorithm
+  (same as fzy).
 
 ==============================================================================
 REMOVED FEATURES                                                 *news-removed*
