+++ runtime/doc/pattern.txt
@@ -1489,6 +1489,9 @@ characters in the search string.  If the search string has multiple words, then
 each word is matched separately.  So the words in the search string can be
 present in any order in a string.
 
+Vim uses the same improved algorithm as the fzy project:
+https://github.com/jhawthorn/fzy
+
 Fuzzy matching assigns a score for each matched string based on the following
 criteria:
     - The number of sequentially matching characters.
