+++ runtime/doc/vimfn.txt
@@ -6520,9 +6520,6 @@ matchfuzzy({list}, {str} [, {dict}])                              *matchfuzzy()*
 				given sequence.
 		    limit	Maximum number of matches in {list} to be
 				returned.  Zero means no limit.
-		    camelcase	Use enhanced camel case scoring making results
-				better suited for completion related to
-				programming languages.  Defaults to v:true.
 
 		If {list} is a list of dictionaries, then the optional {dict}
 		argument supports the following additional items:
