+++ runtime/lua/vim/_meta/vimfn.lua
@@ -5918,9 +5918,6 @@ function vim.fn.matchend(expr, pat, start, count) end
 ---     given sequence.
 ---     limit  Maximum number of matches in {list} to be
 ---     returned.  Zero means no limit.
----     camelcase  Use enhanced camel case scoring making results
----     better suited for completion related to
----     programming languages.  Defaults to v:true.
 ---
 --- If {list} is a list of dictionaries, then the optional {dict}
 --- argument supports the following additional items:
