+++ src/nvim/eval.lua
@@ -7272,9 +7272,6 @@ M.funcs = {
       		given sequence.
           limit	Maximum number of matches in {list} to be
       		returned.  Zero means no limit.
-          camelcase	Use enhanced camel case scoring making results
-      		better suited for completion related to
-      		programming languages.  Defaults to v:true.
 
       If {list} is a list of dictionaries, then the optional {dict}
       argument supports the following additional items:
