+++ src/nvim/fuzzy.h
@@ -1,0 +1,25 @@
+#pragma once
+
+#include <limits.h>
+#include <stdint.h>  // IWYU pragma: keep
+
+#include "nvim/buffer_defs.h"  // IWYU pragma: keep
+#include "nvim/eval/typval_defs.h"  // IWYU pragma: keep
+#include "nvim/garray_defs.h"  // IWYU pragma: keep
+#include "nvim/pos_defs.h"  // IWYU pragma: keep
+#include "nvim/types_defs.h"  // IWYU pragma: keep
+
+enum { FUZZY_MATCH_MAX_LEN = 1024, };  ///< max characters that can be matched
+enum { FUZZY_SCORE_NONE = INT_MIN, };  ///< invalid fuzzy score
+
+/// Fuzzy matched string list item. Used for fuzzy match completion. Items are
+/// usually sorted by "score". The "idx" member is used for stable-sort.
+typedef struct {
+  int idx;
+  char *str;
+  int score;
+} fuzmatch_str_T;
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "fuzzy.h.generated.h"
+#endif
