+++ src/nvim/search.c
@@ -29,8 +29,6 @@
 #include "nvim/file_search.h"
 #include "nvim/fileio.h"
 #include "nvim/fold.h"
-#include "nvim/garray.h"
-#include "nvim/garray_defs.h"
 #include "nvim/getchar.h"
 #include "nvim/gettext_defs.h"
 #include "nvim/globals.h"
@@ -2908,925 +2906,6 @@ void f_searchcount(typval_T *argvars, typval_T *rettv, EvalFuncData fptr)
 the_end:
   restore_last_search_pattern();
   restore_incsearch_state();
-}
-
-/// Fuzzy string matching
-///
-/// Ported from the lib_fts library authored by Forrest Smith.
-/// https://github.com/forrestthewoods/lib_fts/tree/master/code
-///
-/// The following blog describes the fuzzy matching algorithm:
-/// https://www.forrestthewoods.com/blog/reverse_engineering_sublime_texts_fuzzy_match/
-///
-/// Each matching string is assigned a score. The following factors are checked:
-///   - Matched letter
-///   - Unmatched letter
-///   - Consecutively matched letters
-///   - Proximity to start
-///   - Letter following a separator (space, underscore)
-///   - Uppercase letter following lowercase (aka CamelCase)
-///
-/// Matched letters are good. Unmatched letters are bad. Matching near the start
-/// is good. Matching the first letter in the middle of a phrase is good.
-/// Matching the uppercase letters in camel case entries is good.
-///
-/// The score assigned for each factor is explained below.
-/// File paths are different from file names. File extensions may be ignorable.
-/// Single words care about consecutive matches but not separators or camel
-/// case.
-///   Score starts at 100
-///   Matched letter: +0 points
-///   Unmatched letter: -1 point
-///   Consecutive match bonus: +15 points
-///   First letter bonus: +15 points
-///   Separator bonus: +30 points
-///   Camel case bonus: +30 points
-///   Unmatched leading letter: -5 points (max: -15)
-///
-/// There is some nuance to this. Scores don’t have an intrinsic meaning. The
-/// score range isn’t 0 to 100. It’s roughly [50, 150]. Longer words have a
-/// lower minimum score due to unmatched letter penalty. Longer search patterns
-/// have a higher maximum score due to match bonuses.
-///
-/// Separator and camel case bonus is worth a LOT. Consecutive matches are worth
-/// quite a bit.
-///
-/// There is a penalty if you DON’T match the first three letters. Which
-/// effectively rewards matching near the start. However there’s no difference
-/// in matching between the middle and end.
-///
-/// There is not an explicit bonus for an exact match. Unmatched letters receive
-/// a penalty. So shorter strings and closer matches are worth more.
-typedef struct {
-  int idx;  ///< used for stable sort
-  listitem_T *item;
-  int score;
-  list_T *lmatchpos;
-} fuzzyItem_T;
-
-/// bonus for adjacent matches; this is higher than SEPARATOR_BONUS so that
-/// matching a whole word is preferred.
-#define SEQUENTIAL_BONUS 40
-/// bonus if match occurs after a path separator
-#define PATH_SEPARATOR_BONUS 30
-/// bonus if match occurs after a word separator
-#define WORD_SEPARATOR_BONUS 25
-/// bonus if match is uppercase and prev is lower
-#define CAMEL_BONUS 30
-/// bonus if the first letter is matched
-#define FIRST_LETTER_BONUS 15
-/// bonus if exact match
-#define EXACT_MATCH_BONUS 100
-/// bonus if case match when no ignorecase
-#define CASE_MATCH_BONUS 25
-/// penalty applied for every letter in str before the first match
-#define LEADING_LETTER_PENALTY (-5)
-/// maximum penalty for leading letters
-#define MAX_LEADING_LETTER_PENALTY (-15)
-/// penalty for every letter that doesn't match
-#define UNMATCHED_LETTER_PENALTY (-1)
-/// penalty for gap in matching positions (-2 * k)
-#define GAP_PENALTY (-2)
-/// Score for a string that doesn't fuzzy match the pattern
-#define SCORE_NONE (-9999)
-
-#define FUZZY_MATCH_RECURSION_LIMIT 10
-
-/// Compute a score for a fuzzy matched string. The matching character locations
-/// are in "matches".
-static int fuzzy_match_compute_score(const char *const fuzpat, const char *const str,
-                                     const int strSz, const uint32_t *const matches,
-                                     const int numMatches, bool camelcase)
-  FUNC_ATTR_NONNULL_ALL FUNC_ATTR_WARN_UNUSED_RESULT FUNC_ATTR_PURE
-{
-  assert(numMatches > 0);  // suppress clang "result of operation is garbage"
-  const char *p = str;
-  uint32_t sidx = 0;
-  bool is_exact_match = true;
-  const char *const orig_fuzpat = fuzpat - numMatches;
-  const char *curpat = orig_fuzpat;
-  int pat_idx = 0;
-  // Track consecutive camel case matches
-  int consecutive_camel = 0;
-
-  // Initialize score
-  int score = 100;
-
-  // Apply leading letter penalty
-  int penalty = LEADING_LETTER_PENALTY * (int)matches[0];
-  if (penalty < MAX_LEADING_LETTER_PENALTY) {
-    penalty = MAX_LEADING_LETTER_PENALTY;
-  }
-  score += penalty;
-
-  // Apply unmatched penalty
-  const int unmatched = strSz - numMatches;
-  score += UNMATCHED_LETTER_PENALTY * unmatched;
-  // In a long string, not all matches may be found due to the recursion limit.
-  // If at least one match is found, reset the score to a non-negative value.
-  if (score < 0 && numMatches > 0) {
-    score = 0;
-  }
-
-  // Apply ordering bonuses
-  for (int i = 0; i < numMatches; i++) {
-    const uint32_t currIdx = matches[i];
-    bool is_camel = false;
-
-    if (i > 0) {
-      const uint32_t prevIdx = matches[i - 1];
-
-      // Sequential
-      if (currIdx == prevIdx + 1) {
-        score += SEQUENTIAL_BONUS;
-      } else {
-        score += GAP_PENALTY * (int)(currIdx - prevIdx);
-        // Reset consecutive camel count on gap
-        consecutive_camel = 0;
-      }
-    }
-
-    int curr;
-    // Check for bonuses based on neighbor character value
-    if (currIdx > 0) {
-      // Camel case
-      int neighbor = ' ';
-
-      while (sidx < currIdx) {
-        neighbor = utf_ptr2char(p);
-        MB_PTR_ADV(p);
-        sidx++;
-      }
-      curr = utf_ptr2char(p);
-
-      // Enhanced camel case scoring
-      if (camelcase && mb_islower(neighbor) && mb_isupper(curr)) {
-        score += CAMEL_BONUS * 2;  // Double the camel case bonus
-        is_camel = true;
-        consecutive_camel++;
-        // Additional bonus for consecutive camel
-        if (consecutive_camel > 1) {
-          score += CAMEL_BONUS;
-        }
-      } else {
-        consecutive_camel = 0;
-      }
-
-      // Bonus if the match follows a separator character
-      if (neighbor == '/' || neighbor == '\\') {
-        score += PATH_SEPARATOR_BONUS;
-      } else if (neighbor == ' ' || neighbor == '_') {
-        score += WORD_SEPARATOR_BONUS;
-      }
-    } else {
-      // First letter
-      score += FIRST_LETTER_BONUS;
-      curr = utf_ptr2char(p);
-    }
-
-    // Case matching bonus
-    if (mb_isalpha(curr)) {
-      while (pat_idx < i && *curpat) {
-        MB_PTR_ADV(curpat);
-        pat_idx++;
-      }
-
-      if (curr == utf_ptr2char(curpat)) {
-        score += CASE_MATCH_BONUS;
-        // Extra bonus for exact case match in camel
-        if (is_camel) {
-          score += CASE_MATCH_BONUS / 2;
-        }
-      }
-    }
-
-    // Check exact match condition
-    if (currIdx != (uint32_t)i) {
-      is_exact_match = false;
-    }
-  }
-
-  // Boost score for exact matches
-  if (is_exact_match && numMatches == strSz) {
-    score += EXACT_MATCH_BONUS;
-  }
-
-  return score;
-}
-
-/// Perform a recursive search for fuzzy matching "fuzpat" in "str".
-/// @return the number of matching characters.
-static int fuzzy_match_recursive(const char *fuzpat, const char *str, uint32_t strIdx,
-                                 int *const outScore, const char *const strBegin, const int strLen,
-                                 const uint32_t *const srcMatches, uint32_t *const matches,
-                                 const int maxMatches, int nextMatch, int *const recursionCount,
-                                 bool camelcase)
-  FUNC_ATTR_NONNULL_ARG(1, 2, 4, 5, 8, 11) FUNC_ATTR_WARN_UNUSED_RESULT
-{
-  // Recursion params
-  bool recursiveMatch = false;
-  uint32_t bestRecursiveMatches[MAX_FUZZY_MATCHES];
-  int bestRecursiveScore = 0;
-
-  // Count recursions
-  (*recursionCount)++;
-  if (*recursionCount >= FUZZY_MATCH_RECURSION_LIMIT) {
-    return 0;
-  }
-
-  // Detect end of strings
-  if (*fuzpat == NUL || *str == NUL) {
-    return 0;
-  }
-
-  // Loop through fuzpat and str looking for a match
-  bool first_match = true;
-  while (*fuzpat != NUL && *str != NUL) {
-    const int c1 = utf_ptr2char(fuzpat);
-    const int c2 = utf_ptr2char(str);
-
-    // Found match
-    if (mb_tolower(c1) == mb_tolower(c2)) {
-      // Supplied matches buffer was too short
-      if (nextMatch >= maxMatches) {
-        return 0;
-      }
-
-      int recursiveScore = 0;
-      uint32_t recursiveMatches[MAX_FUZZY_MATCHES];
-      CLEAR_FIELD(recursiveMatches);
-
-      // "Copy-on-Write" srcMatches into matches
-      if (first_match && srcMatches != NULL) {
-        memcpy(matches, srcMatches, (size_t)nextMatch * sizeof(srcMatches[0]));
-        first_match = false;
-      }
-
-      // Recursive call that "skips" this match
-      const char *const next_char = str + utfc_ptr2len(str);
-      if (fuzzy_match_recursive(fuzpat, next_char, strIdx + 1, &recursiveScore, strBegin, strLen,
-                                matches, recursiveMatches,
-                                sizeof(recursiveMatches) / sizeof(recursiveMatches[0]), nextMatch,
-                                recursionCount, camelcase)) {
-        // Pick best recursive score
-        if (!recursiveMatch || recursiveScore > bestRecursiveScore) {
-          memcpy(bestRecursiveMatches, recursiveMatches,
-                 MAX_FUZZY_MATCHES * sizeof(recursiveMatches[0]));
-          bestRecursiveScore = recursiveScore;
-        }
-        recursiveMatch = true;
-      }
-
-      // Advance
-      matches[nextMatch++] = strIdx;
-      MB_PTR_ADV(fuzpat);
-    }
-    MB_PTR_ADV(str);
-    strIdx++;
-  }
-
-  // Determine if full fuzpat was matched
-  const bool matched = *fuzpat == NUL;
-
-  // Calculate score
-  if (matched) {
-    *outScore = fuzzy_match_compute_score(fuzpat, strBegin, strLen, matches, nextMatch, camelcase);
-  }
-
-  // Return best result
-  if (recursiveMatch && (!matched || bestRecursiveScore > *outScore)) {
-    // Recursive score is better than "this"
-    memcpy(matches, bestRecursiveMatches, (size_t)maxMatches * sizeof(matches[0]));
-    *outScore = bestRecursiveScore;
-    return nextMatch;
-  } else if (matched) {
-    return nextMatch;  // "this" score is better than recursive
-  }
-
-  return 0;  // no match
-}
-
-/// fuzzy_match()
-///
-/// Performs exhaustive search via recursion to find all possible matches and
-/// match with highest score.
-/// Scores values have no intrinsic meaning.  Possible score range is not
-/// normalized and varies with pattern.
-/// Recursion is limited internally (default=10) to prevent degenerate cases
-/// (pat_arg="aaaaaa" str="aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa").
-/// Patterns are limited to MAX_FUZZY_MATCHES characters.
-///
-/// @return true if "pat_arg" matches "str". Also returns the match score in
-/// "outScore" and the matching character positions in "matches".
-bool fuzzy_match(char *const str, const char *const pat_arg, const bool matchseq,
-                 int *const outScore, uint32_t *const matches, const int maxMatches, bool camelcase)
-  FUNC_ATTR_NONNULL_ALL
-{
-  const int len = mb_charlen(str);
-  bool complete = false;
-  int numMatches = 0;
-
-  *outScore = 0;
-
-  char *const save_pat = xstrdup(pat_arg);
-  char *pat = save_pat;
-  char *p = pat;
-
-  // Try matching each word in "pat_arg" in "str"
-  while (true) {
-    if (matchseq) {
-      complete = true;
-    } else {
-      // Extract one word from the pattern (separated by space)
-      p = skipwhite(p);
-      if (*p == NUL) {
-        break;
-      }
-      pat = p;
-      while (*p != NUL && !ascii_iswhite(utf_ptr2char(p))) {
-        MB_PTR_ADV(p);
-      }
-      if (*p == NUL) {  // processed all the words
-        complete = true;
-      }
-      *p = NUL;
-    }
-
-    int score = 0;
-    int recursionCount = 0;
-    const int matchCount
-      = fuzzy_match_recursive(pat, str, 0, &score, str, len, NULL,
-                              matches + numMatches,
-                              maxMatches - numMatches, 0, &recursionCount, camelcase);
-    if (matchCount == 0) {
-      numMatches = 0;
-      break;
-    }
-
-    // Accumulate the match score and the number of matches
-    *outScore += score;
-    numMatches += matchCount;
-
-    if (complete) {
-      break;
-    }
-
-    // try matching the next word
-    p++;
-  }
-
-  xfree(save_pat);
-  return numMatches != 0;
-}
-
-/// Sort the fuzzy matches in the descending order of the match score.
-/// For items with same score, retain the order using the index (stable sort)
-static int fuzzy_match_item_compare(const void *const s1, const void *const s2)
-  FUNC_ATTR_NONNULL_ALL FUNC_ATTR_WARN_UNUSED_RESULT FUNC_ATTR_PURE
-{
-  const int v1 = ((const fuzzyItem_T *)s1)->score;
-  const int v2 = ((const fuzzyItem_T *)s2)->score;
-  const int idx1 = ((const fuzzyItem_T *)s1)->idx;
-  const int idx2 = ((const fuzzyItem_T *)s2)->idx;
-
-  if (v1 == v2) {
-    return idx1 == idx2 ? 0 : idx1 > idx2 ? 1 : -1;
-  }
-  return v1 > v2 ? -1 : 1;
-}
-
-/// Fuzzy search the string "str" in a list of "items" and return the matching
-/// strings in "fmatchlist".
-/// If "matchseq" is true, then for multi-word search strings, match all the
-/// words in sequence.
-/// If "items" is a list of strings, then search for "str" in the list.
-/// If "items" is a list of dicts, then either use "key" to lookup the string
-/// for each item or use "item_cb" Funcref function to get the string.
-/// If "retmatchpos" is true, then return a list of positions where "str"
-/// matches for each item.
-static void fuzzy_match_in_list(list_T *const l, char *const str, const bool matchseq,
-                                const char *const key, Callback *const item_cb,
-                                const bool retmatchpos, list_T *const fmatchlist,
-                                const int max_matches, bool camelcase)
-  FUNC_ATTR_NONNULL_ARG(2, 5, 7)
-{
-  int len = tv_list_len(l);
-  if (len == 0) {
-    return;
-  }
-  if (max_matches > 0 && len > max_matches) {
-    len = max_matches;
-  }
-
-  fuzzyItem_T *const items = xcalloc((size_t)len, sizeof(fuzzyItem_T));
-  int match_count = 0;
-  uint32_t matches[MAX_FUZZY_MATCHES];
-
-  // For all the string items in items, get the fuzzy matching score
-  TV_LIST_ITER(l, li, {
-    if (max_matches > 0 && match_count >= max_matches) {
-      break;
-    }
-
-    char *itemstr = NULL;
-    typval_T rettv;
-    rettv.v_type = VAR_UNKNOWN;
-    const typval_T *const tv = TV_LIST_ITEM_TV(li);
-    if (tv->v_type == VAR_STRING) {  // list of strings
-      itemstr = tv->vval.v_string;
-    } else if (tv->v_type == VAR_DICT && (key != NULL || item_cb->type != kCallbackNone)) {
-      // For a dict, either use the specified key to lookup the string or
-      // use the specified callback function to get the string.
-      if (key != NULL) {
-        itemstr = tv_dict_get_string(tv->vval.v_dict, key, false);
-      } else {
-        typval_T argv[2];
-
-        // Invoke the supplied callback (if any) to get the dict item
-        tv->vval.v_dict->dv_refcount++;
-        argv[0].v_type = VAR_DICT;
-        argv[0].vval.v_dict = tv->vval.v_dict;
-        argv[1].v_type = VAR_UNKNOWN;
-        if (callback_call(item_cb, 1, argv, &rettv)) {
-          if (rettv.v_type == VAR_STRING) {
-            itemstr = rettv.vval.v_string;
-          }
-        }
-        tv_dict_unref(tv->vval.v_dict);
-      }
-    }
-
-    int score;
-    if (itemstr != NULL && fuzzy_match(itemstr, str, matchseq, &score, matches,
-                                       MAX_FUZZY_MATCHES, camelcase)) {
-      items[match_count].idx = (int)match_count;
-      items[match_count].item = li;
-      items[match_count].score = score;
-
-      // Copy the list of matching positions in itemstr to a list, if
-      // "retmatchpos" is set.
-      if (retmatchpos) {
-        items[match_count].lmatchpos = tv_list_alloc(kListLenMayKnow);
-        int j = 0;
-        const char *p = str;
-        while (*p != NUL) {
-          if (!ascii_iswhite(utf_ptr2char(p)) || matchseq) {
-            tv_list_append_number(items[match_count].lmatchpos, matches[j]);
-            j++;
-          }
-          MB_PTR_ADV(p);
-        }
-      }
-      match_count++;
-    }
-    tv_clear(&rettv);
-  });
-
-  if (match_count > 0) {
-    // Sort the list by the descending order of the match score
-    qsort(items, (size_t)match_count, sizeof(fuzzyItem_T), fuzzy_match_item_compare);
-
-    // For matchfuzzy(), return a list of matched strings.
-    //          ['str1', 'str2', 'str3']
-    // For matchfuzzypos(), return a list with three items.
-    // The first item is a list of matched strings. The second item
-    // is a list of lists where each list item is a list of matched
-    // character positions. The third item is a list of matching scores.
-    //      [['str1', 'str2', 'str3'], [[1, 3], [1, 3], [1, 3]]]
-    list_T *retlist;
-    if (retmatchpos) {
-      const listitem_T *const li = tv_list_find(fmatchlist, 0);
-      assert(li != NULL && TV_LIST_ITEM_TV(li)->vval.v_list != NULL);
-      retlist = TV_LIST_ITEM_TV(li)->vval.v_list;
-    } else {
-      retlist = fmatchlist;
-    }
-
-    // Copy the matching strings with a valid score to the return list
-    for (int i = 0; i < match_count; i++) {
-      if (items[i].score == SCORE_NONE) {
-        break;
-      }
-      tv_list_append_tv(retlist, TV_LIST_ITEM_TV(items[i].item));
-    }
-
-    // next copy the list of matching positions
-    if (retmatchpos) {
-      const listitem_T *li = tv_list_find(fmatchlist, -2);
-      assert(li != NULL && TV_LIST_ITEM_TV(li)->vval.v_list != NULL);
-      retlist = TV_LIST_ITEM_TV(li)->vval.v_list;
-
-      for (int i = 0; i < match_count; i++) {
-        if (items[i].score == SCORE_NONE) {
-          break;
-        }
-        tv_list_append_list(retlist, items[i].lmatchpos);
-      }
-
-      // copy the matching scores
-      li = tv_list_find(fmatchlist, -1);
-      assert(li != NULL && TV_LIST_ITEM_TV(li)->vval.v_list != NULL);
-      retlist = TV_LIST_ITEM_TV(li)->vval.v_list;
-      for (int i = 0; i < match_count; i++) {
-        if (items[i].score == SCORE_NONE) {
-          break;
-        }
-        tv_list_append_number(retlist, items[i].score);
-      }
-    }
-  }
-  xfree(items);
-}
-
-/// Do fuzzy matching. Returns the list of matched strings in "rettv".
-/// If "retmatchpos" is true, also returns the matching character positions.
-static void do_fuzzymatch(const typval_T *const argvars, typval_T *const rettv,
-                          const bool retmatchpos)
-  FUNC_ATTR_NONNULL_ALL
-{
-  // validate and get the arguments
-  if (argvars[0].v_type != VAR_LIST || argvars[0].vval.v_list == NULL) {
-    semsg(_(e_listarg), retmatchpos ? "matchfuzzypos()" : "matchfuzzy()");
-    return;
-  }
-  if (argvars[1].v_type != VAR_STRING || argvars[1].vval.v_string == NULL) {
-    semsg(_(e_invarg2), tv_get_string(&argvars[1]));
-    return;
-  }
-
-  Callback cb = CALLBACK_NONE;
-  const char *key = NULL;
-  bool matchseq = false;
-  int max_matches = 0;
-  bool camelcase = true;
-  if (argvars[2].v_type != VAR_UNKNOWN) {
-    if (tv_check_for_nonnull_dict_arg(argvars, 2) == FAIL) {
-      return;
-    }
-
-    // To search a dict, either a callback function or a key can be
-    // specified.
-    dict_T *const d = argvars[2].vval.v_dict;
-    const dictitem_T *di;
-    if ((di = tv_dict_find(d, "key", -1)) != NULL) {
-      if (di->di_tv.v_type != VAR_STRING || di->di_tv.vval.v_string == NULL
-          || *di->di_tv.vval.v_string == NUL) {
-        semsg(_(e_invargNval), "key", tv_get_string(&di->di_tv));
-        return;
-      }
-      key = tv_get_string(&di->di_tv);
-    } else if (!tv_dict_get_callback(d, "text_cb", -1, &cb)) {
-      semsg(_(e_invargval), "text_cb");
-      return;
-    }
-
-    if ((di = tv_dict_find(d, "limit", -1)) != NULL) {
-      if (di->di_tv.v_type != VAR_NUMBER) {
-        semsg(_(e_invargval), "limit");
-        return;
-      }
-      max_matches = (int)tv_get_number_chk(&di->di_tv, NULL);
-    }
-
-    if ((di = tv_dict_find(d, "camelcase", -1)) != NULL) {
-      if (di->di_tv.v_type != VAR_BOOL) {
-        semsg(_(e_invargval), "camelcase");
-        return;
-      }
-      camelcase = tv_get_bool_chk(&di->di_tv, NULL);
-    }
-
-    if (tv_dict_find(d, "matchseq", -1) != NULL) {
-      matchseq = true;
-    }
-  }
-
-  // get the fuzzy matches
-  tv_list_alloc_ret(rettv, retmatchpos ? 3 : kListLenUnknown);
-  if (retmatchpos) {
-    // For matchfuzzypos(), a list with three items are returned. First
-    // item is a list of matching strings, the second item is a list of
-    // lists with matching positions within each string and the third item
-    // is the list of scores of the matches.
-    tv_list_append_list(rettv->vval.v_list, tv_list_alloc(kListLenUnknown));
-    tv_list_append_list(rettv->vval.v_list, tv_list_alloc(kListLenUnknown));
-    tv_list_append_list(rettv->vval.v_list, tv_list_alloc(kListLenUnknown));
-  }
-
-  fuzzy_match_in_list(argvars[0].vval.v_list,
-                      (char *)tv_get_string(&argvars[1]), matchseq, key,
-                      &cb, retmatchpos, rettv->vval.v_list, max_matches, camelcase);
-  callback_free(&cb);
-}
-
-/// "matchfuzzy()" function
-void f_matchfuzzy(typval_T *argvars, typval_T *rettv, EvalFuncData fptr)
-{
-  do_fuzzymatch(argvars, rettv, false);
-}
-
-/// "matchfuzzypos()" function
-void f_matchfuzzypos(typval_T *argvars, typval_T *rettv, EvalFuncData fptr)
-{
-  do_fuzzymatch(argvars, rettv, true);
-}
-
-/// Same as fuzzy_match_item_compare() except for use with a string match
-static int fuzzy_match_str_compare(const void *const s1, const void *const s2)
-  FUNC_ATTR_WARN_UNUSED_RESULT FUNC_ATTR_NONNULL_ALL FUNC_ATTR_PURE
-{
-  const int v1 = ((fuzmatch_str_T *)s1)->score;
-  const int v2 = ((fuzmatch_str_T *)s2)->score;
-  const int idx1 = ((fuzmatch_str_T *)s1)->idx;
-  const int idx2 = ((fuzmatch_str_T *)s2)->idx;
-
-  if (v1 == v2) {
-    return idx1 == idx2 ? 0 : idx1 > idx2 ? 1 : -1;
-  } else {
-    return v1 > v2 ? -1 : 1;
-  }
-}
-
-/// Sort fuzzy matches by score
-static void fuzzy_match_str_sort(fuzmatch_str_T *const fm, const int sz)
-  FUNC_ATTR_NONNULL_ALL
-{
-  // Sort the list by the descending order of the match score
-  qsort(fm, (size_t)sz, sizeof(fuzmatch_str_T), fuzzy_match_str_compare);
-}
-
-/// Same as fuzzy_match_item_compare() except for use with a function name
-/// string match. <SNR> functions should be sorted to the end.
-static int fuzzy_match_func_compare(const void *const s1, const void *const s2)
-  FUNC_ATTR_WARN_UNUSED_RESULT FUNC_ATTR_NONNULL_ALL FUNC_ATTR_PURE
-{
-  const int v1 = ((fuzmatch_str_T *)s1)->score;
-  const int v2 = ((fuzmatch_str_T *)s2)->score;
-  const int idx1 = ((fuzmatch_str_T *)s1)->idx;
-  const int idx2 = ((fuzmatch_str_T *)s2)->idx;
-  const char *const str1 = ((fuzmatch_str_T *)s1)->str;
-  const char *const str2 = ((fuzmatch_str_T *)s2)->str;
-
-  if (*str1 != '<' && *str2 == '<') {
-    return -1;
-  }
-  if (*str1 == '<' && *str2 != '<') {
-    return 1;
-  }
-  if (v1 == v2) {
-    return idx1 == idx2 ? 0 : idx1 > idx2 ? 1 : -1;
-  }
-  return v1 > v2 ? -1 : 1;
-}
-
-/// Sort fuzzy matches of function names by score.
-/// <SNR> functions should be sorted to the end.
-static void fuzzy_match_func_sort(fuzmatch_str_T *const fm, const int sz)
-  FUNC_ATTR_NONNULL_ALL
-{
-  // Sort the list by the descending order of the match score
-  qsort(fm, (size_t)sz, sizeof(fuzmatch_str_T), fuzzy_match_func_compare);
-}
-
-/// Fuzzy match "pat" in "str".
-/// @returns 0 if there is no match. Otherwise, returns the match score.
-int fuzzy_match_str(char *const str, const char *const pat)
-  FUNC_ATTR_WARN_UNUSED_RESULT
-{
-  if (str == NULL || pat == NULL) {
-    return 0;
-  }
-
-  int score = 0;
-  uint32_t matchpos[MAX_FUZZY_MATCHES];
-  fuzzy_match(str, pat, true, &score, matchpos, sizeof(matchpos) / sizeof(matchpos[0]), true);
-
-  return score;
-}
-
-/// Fuzzy match the position of string "pat" in string "str".
-/// @returns a dynamic array of matching positions. If there is no match, returns NULL.
-garray_T *fuzzy_match_str_with_pos(char *const str, const char *const pat)
-{
-  if (str == NULL || pat == NULL) {
-    return NULL;
-  }
-
-  garray_T *match_positions = xmalloc(sizeof(garray_T));
-  ga_init(match_positions, sizeof(uint32_t), 10);
-
-  unsigned matches[MAX_FUZZY_MATCHES];
-  int score = 0;
-  if (!fuzzy_match(str, pat, false, &score, matches, MAX_FUZZY_MATCHES, true)
-      || score == 0) {
-    ga_clear(match_positions);
-    xfree(match_positions);
-    return NULL;
-  }
-
-  int j = 0;
-  for (const char *p = pat; *p != NUL; MB_PTR_ADV(p)) {
-    if (!ascii_iswhite(utf_ptr2char(p))) {
-      GA_APPEND(uint32_t, match_positions, matches[j]);
-      j++;
-    }
-  }
-
-  return match_positions;
-}
-
-/// This function splits the line pointed to by `*ptr` into words and performs
-/// a fuzzy match for the pattern `pat` on each word. It iterates through the
-/// line, moving `*ptr` to the start of each word during the process.
-///
-/// If a match is found:
-/// - `*ptr` points to the start of the matched word.
-/// - `*len` is set to the length of the matched word.
-/// - `*score` contains the match score.
-///
-/// If no match is found, `*ptr` is updated to the end of the line.
-bool fuzzy_match_str_in_line(char **ptr, char *pat, int *len, pos_T *current_pos, int *score)
-{
-  char *str = *ptr;
-  char *strBegin = str;
-  char *end = NULL;
-  char *start = NULL;
-  bool found = false;
-
-  if (str == NULL || pat == NULL) {
-    return found;
-  }
-  char *line_end = find_line_end(str);
-
-  while (str < line_end) {
-    // Skip non-word characters
-    start = find_word_start(str);
-    if (*start == NUL) {
-      break;
-    }
-    end = find_word_end(start);
-
-    // Extract the word from start to end
-    char save_end = *end;
-    *end = NUL;
-
-    // Perform fuzzy match
-    *score = fuzzy_match_str(start, pat);
-    *end = save_end;
-
-    if (*score > 0) {
-      *len = (int)(end - start);
-      found = true;
-      *ptr = start;
-      if (current_pos) {
-        current_pos->col += (int)(end - strBegin);
-      }
-      break;
-    }
-
-    // Move to the end of the current word for the next iteration
-    str = end;
-    // Ensure we continue searching after the current word
-    while (*str != NUL && !vim_iswordp(str)) {
-      MB_PTR_ADV(str);
-    }
-  }
-
-  if (!found) {
-    *ptr = line_end;
-  }
-
-  return found;
-}
-
-/// Search for the next fuzzy match in the specified buffer.
-/// This function attempts to find the next occurrence of the given pattern
-/// in the buffer, starting from the current position. It handles line wrapping
-/// and direction of search.
-///
-/// Return true if a match is found, otherwise false.
-bool search_for_fuzzy_match(buf_T *buf, pos_T *pos, char *pattern, int dir, pos_T *start_pos,
-                            int *len, char **ptr, int *score)
-{
-  pos_T current_pos = *pos;
-  pos_T circly_end;
-  bool found_new_match = false;
-  bool looped_around = false;
-
-  bool whole_line = ctrl_x_mode_whole_line();
-
-  if (buf == curbuf) {
-    circly_end = *start_pos;
-  } else {
-    circly_end.lnum = buf->b_ml.ml_line_count;
-    circly_end.col = 0;
-    circly_end.coladd = 0;
-  }
-
-  if (whole_line && start_pos->lnum != pos->lnum) {
-    current_pos.lnum += dir;
-  }
-
-  while (true) {
-    // Check if looped around and back to start position
-    if (looped_around && equalpos(current_pos, circly_end)) {
-      break;
-    }
-
-    // Ensure current_pos is valid
-    if (current_pos.lnum >= 1 && current_pos.lnum <= buf->b_ml.ml_line_count) {
-      // Get the current line buffer
-      *ptr = ml_get_buf(buf, current_pos.lnum);
-      if (!whole_line) {
-        *ptr += current_pos.col;
-      }
-
-      // If ptr is end of line is reached, move to next line
-      // or previous line based on direction
-      if (*ptr != NULL && **ptr != NUL) {
-        if (!whole_line) {
-          // Try to find a fuzzy match in the current line starting
-          // from current position
-          found_new_match = fuzzy_match_str_in_line(ptr, pattern,
-                                                    len, &current_pos, score);
-          if (found_new_match) {
-            *pos = current_pos;
-            break;
-          } else if (looped_around && current_pos.lnum == circly_end.lnum) {
-            break;
-          }
-        } else {
-          if (fuzzy_match_str(*ptr, pattern) > 0) {
-            found_new_match = true;
-            *pos = current_pos;
-            *len = ml_get_buf_len(buf, current_pos.lnum);
-            break;
-          }
-        }
-      }
-    }
-
-    // Move to the next line or previous line based on direction
-    if (dir == FORWARD) {
-      if (++current_pos.lnum > buf->b_ml.ml_line_count) {
-        if (p_ws) {
-          current_pos.lnum = 1;
-          looped_around = true;
-        } else {
-          break;
-        }
-      }
-    } else {
-      if (--current_pos.lnum < 1) {
-        if (p_ws) {
-          current_pos.lnum = buf->b_ml.ml_line_count;
-          looped_around = true;
-        } else {
-          break;
-        }
-      }
-    }
-    current_pos.col = 0;
-  }
-
-  return found_new_match;
-}
-
-/// Copy a list of fuzzy matches into a string list after sorting the matches by
-/// the fuzzy score. Frees the memory allocated for "fuzmatch".
-void fuzzymatches_to_strmatches(fuzmatch_str_T *const fuzmatch, char ***const matches,
-                                const int count, const bool funcsort)
-  FUNC_ATTR_NONNULL_ARG(2)
-{
-  if (count <= 0) {
-    return;
-  }
-
-  *matches = xmalloc((size_t)count * sizeof(char *));
-
-  // Sort the list by the descending order of the match score
-  if (funcsort) {
-    fuzzy_match_func_sort(fuzmatch, count);
-  } else {
-    fuzzy_match_str_sort(fuzmatch, count);
-  }
-
-  for (int i = 0; i < count; i++) {
-    (*matches)[i] = fuzmatch[i].str;
-  }
-  xfree(fuzmatch);
-}
-
-/// Free a list of fuzzy string matches.
-void fuzmatch_str_free(fuzmatch_str_T *const fuzmatch, int count)
-{
-  if (count <= 0 || fuzmatch == NULL) {
-    return;
-  }
-  while (count--) {
-    xfree(fuzmatch[count].str);
-  }
-  xfree(fuzmatch);
 }
 
 /// Get line "lnum" and copy it into "buf[LSIZE]".
