+++ src/nvim/search.h
@@ -67,11 +67,6 @@ enum { SEARCH_STAT_DEF_TIMEOUT = 40, };
 // '[>9999/>9999]': 13 + 1 (NUL)
 enum { SEARCH_STAT_BUF_LEN = 16, };
 
-enum {
-  /// Maximum number of characters that can be fuzzy matched
-  MAX_FUZZY_MATCHES = 256,
-};
-
 /// Structure containing offset definition for the last search pattern
 ///
 /// @note Only offset for the last search pattern is used, not for the last
@@ -111,14 +106,6 @@ typedef struct {
   // 2: max count exceeded
   int last_maxcount;  // the max count of the last search
 } searchstat_T;
-
-/// Fuzzy matched string list item. Used for fuzzy match completion. Items are
-/// usually sorted by "score". The "idx" member is used for stable-sort.
-typedef struct {
-  int idx;
-  char *str;
-  int score;
-} fuzmatch_str_T;
 
 #ifdef INCLUDE_GENERATED_DECLARATIONS
 # include "search.h.generated.h"
